# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Petagree API",
        "description": "Api",
        "termsOfService": "",
        "contact": {
            "email": "contact@mysite.com"
        },
        "license": {
            "name": "Private License",
            "url": "URL to the license"
        },
        "version": "1.0.0"
    },
    "host": "api.petagree/api",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
        "/association": {
            "put": {
                "tags": [
                    "Association"
                ],
                "description": "Edit all association information",
                "operationId": "putAssociation",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name of the association.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "suburb_id",
                        "in": "query",
                        "description": "Location of the association.",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "state_id",
                        "in": "query",
                        "description": "Location of the association.",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "breed",
                        "in": "query",
                        "description": "Dog breed of the association.",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "about",
                        "in": "query",
                        "description": "General information of the association.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/editAssociation"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity."
                    },
                    "401": {
                        "description": "Failed to authenticate because of bad credentials or an invalid authorization header.<br>-----------------------------------------------------------------------<br>Wrong number of segments.<br>-----------------------------------------------------------------------<br>Token Signature could not be verified.<br>-----------------------------------------------------------------------<br>Could not decode token. The token is an invalid JWS.<br>-----------------------------------------------------------------------<br>Token has expired.<br>-----------------------------------------------------------------------<br>The token has been blacklisted<br>-----------------------------------------------------------------------Unable to authenticate with invalid token.<br>------------------------------------<br>Only association profiles can access this resource."
                    }
                }
            },
            "post": {
                "tags": [
                    "Association"
                ],
                "description": "Request creation of new association",
                "operationId": "postAssociation",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name of the association.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "first_name",
                        "in": "query",
                        "description": "First name of the key representative (profile owner).",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "last_name",
                        "in": "query",
                        "description": "Last name of the key representative (profile owner).",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email of the association",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "contact_number",
                        "in": "query",
                        "description": "Telephone number for contact.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "suburb_id",
                        "in": "query",
                        "description": "Location of the association.",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "state_id",
                        "in": "query",
                        "description": "Location of the association.",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "201": {
                        "description": ""
                    },
                    "422": {
                        "description": "Unprocessable entity."
                    }
                }
            },
            "patch": {
                "tags": [
                    "Association"
                ],
                "description": "Edit some of the association information",
                "operationId": "patchAssociation",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name of the association.",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "suburb_id",
                        "in": "query",
                        "description": "Location of the association.",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "state_id",
                        "in": "query",
                        "description": "Location of the association.",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "breed",
                        "in": "query",
                        "description": "Dog breed of the association.",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "about",
                        "in": "query",
                        "description": "General information of the association.",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/editAssociation"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity."
                    },
                    "401": {
                        "description": "Failed to authenticate because of bad credentials or an invalid authorization header.<br>-----------------------------------------------------------------------<br>Wrong number of segments.<br>-----------------------------------------------------------------------<br>Token Signature could not be verified.<br>-----------------------------------------------------------------------<br>Could not decode token. The token is an invalid JWS.<br>-----------------------------------------------------------------------<br>Token has expired.<br>-----------------------------------------------------------------------<br>The token has been blacklisted<br>-----------------------------------------------------------------------Unable to authenticate with invalid token.<br>------------------------------------<br>Only association profiles can access this resource."
                    }
                }
            }
        },
...
        "/association/members": {
            "post": {
                "tags": [
                    "Association"
                ],
                "description": "Add association members.",
                "operationId": "postAssociationMembers",
                "parameters": [
                    {
                        "name": "suburb_id",
                        "in": "query",
                        "description": "State where member lives.",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "state_id",
                        "in": "query",
                        "description": "Location of the association.",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "member_email",
                        "in": "query",
                        "description": "Mail of the breeder (new member).",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "422": {
                        "description": "Unprocessable Entity."
                    },
                    "403": {
                        "description": "Invite is allready sent to this user!"
                    },
                    "401": {
                        "description": "Failed to authenticate because of bad credentials or an invalid authorization header.<br>-----------------------------------------------------------------------<br>Wrong number of segments.<br>-----------------------------------------------------------------------<br>Token Signature could not be verified.<br>-----------------------------------------------------------------------<br>Could not decode token. The token is an invalid JWS.<br>-----------------------------------------------------------------------<br>Token has expired.<br>-----------------------------------------------------------------------<br>The token has been blacklisted<br>-----------------------------------------------------------------------Unable to authenticate with invalid token.<br>------------------------------------<br>Only association profiles can access this resource."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Association"
                ],
                "description": "Delete association members.",
                "operationId": "deleteAssociationMembers",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Delete member with given ID.",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "404": {
                        "description": "User not found"
                    },
                    "422": {
                        "description": "Unprocessable Entity."
                    },
                    "401": {
                        "description": "Failed to authenticate because of bad credentials or an invalid authorization header.<br>-----------------------------------------------------------------------<br>Wrong number of segments.<br>-----------------------------------------------------------------------<br>Token Signature could not be verified.<br>-----------------------------------------------------------------------<br>Could not decode token. The token is an invalid JWS.<br>-----------------------------------------------------------------------<br>Token has expired.<br>-----------------------------------------------------------------------<br>The token has been blacklisted<br>-----------------------------------------------------------------------Unable to authenticate with invalid token.<br>------------------------------------<br>Only association profiles can access this resource."
                    }
                }
            }
        },
        "/association/with": {
            "get": {
                "tags": [
                    "Association"
                ],
                "description": "Get association relations that are manually specified.",
                "operationId": "getAssociationWith",
                "parameters": [
                    {
                        "name": "relations",
                        "in": "query",
                        "description": "Names of relations that user wants to get information from. Relations avaliable : users,breed,suburb,state,key_members,members",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/associationWith"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity."
                    },
                    "401": {
                        "description": "Failed to authenticate because of bad credentials or an invalid authorization header.<br>-----------------------------------------------------------------------<br>Wrong number of segments.<br>-----------------------------------------------------------------------<br>Token Signature could not be verified.<br>-----------------------------------------------------------------------<br>Could not decode token. The token is an invalid JWS.<br>-----------------------------------------------------------------------<br>Token has expired.<br>-----------------------------------------------------------------------<br>The token has been blacklisted<br>-----------------------------------------------------------------------Unable to authenticate with invalid token.<br>------------------------------------<br>Only association profiles can access this resource."
                    }
                }
            }
        },
        "/association/images": {
            "post": {
                "tags": [
                    "Association"
                ],
                "description": "Post association images.",
                "operationId": "postAssociationImages",
                "parameters": [
                    {
                        "name": "avatar",
                        "in": "query",
                        "description": "Avatar of the association.",
                        "required": true,
                        "type": "file"
                    },
                    {
                        "name": "images",
                        "in": "query",
                        "description": "Images of the association.",
                        "required": true,
                        "type": "array",
                        "items": "file"
                    },
                    {
                        "name": "banner_image",
                        "in": "query",
                        "description": "Banner image of the association.",
                        "required": true,
                        "type": "file"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Size of images is too big"
                    },
                    "422": {
                        "description": "Unprocessable Entity."
                    },
                    "401": {
                        "description": "Failed to authenticate because of bad credentials or an invalid authorization header.<br>-----------------------------------------------------------------------<br>Wrong number of segments.<br>-----------------------------------------------------------------------<br>Token Signature could not be verified.<br>-----------------------------------------------------------------------<br>Could not decode token. The token is an invalid JWS.<br>-----------------------------------------------------------------------<br>Token has expired.<br>-----------------------------------------------------------------------<br>The token has been blacklisted<br>-----------------------------------------------------------------------Unable to authenticate with invalid token.<br>------------------------------------<br>Only association profiles can access this resource."
                    }
                }
            }
        },
...
    },
    "definitions": {
        "editAssociation": {
            "properties": {
                "id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "suburb_id": {
                    "type": "integer"
                },
                "state_id": {
                    "type": "integer"
                },
                "avatar": {
                    "type": "string"
                },
                "breed": {
                    "type": "integer"
                },
                "about": {
                    "type": "string"
                },
                "banner_image": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                }
            }
        },
...
        "associationWith": {
            "properties": {
                "id": {
                    "description": "Class AssociationModel",
                    "type": "integer"
                },
                "user_id": {
                    "description": "Class AssociationModel",
                    "type": "integer"
                },
                "name": {
                    "description": "Class AssociationModel",
                    "type": "string"
                },
                "avatar": {
                    "description": " Avatar",
                    "type": "string"
                },
                "about": {
                    "description": " General information of the association",
                    "type": "string"
                },
                "banner_image": {
                    "description": " Banner image",
                    "type": "string"
                },
                "slug": {
                    "description": "Slug",
                    "type": "string"
                },
                "breed": {
                    "description": "Breed relation",
                    "type": "array",
                    "items": "string"
                },
                "suburb": {
                    "description": "Suburb relation",
                    "type": "array",
                    "items": "string"
                },
                "state": {
                    "description": "State relation",
                    "type": "array",
                    "items": "string"
                },
                "members": {
                    "description": "Members relation",
                    "type": "array",
                    "items": "string"
                },
                "key_members": {
                    "description": "Key members relation",
                    "type": "array",
                    "items": "string"
                }
            }
        },
...
    },
...
}